+++ tests/unit/test_database.py
@@ -300,6 +300,7 @@ def test_database_loaded(tdb_with_data):
         {"climate_model": "cmodel_a"},
         {"climate_model": "cmodel_b"},
         {"climate_model": "cmodel_a", "variable": "variable"},
+        {"climate_model": ["cmodel_a", "cmodel_c"]},
     ],
 )
 def test_database_load_data(tdb_with_data, start_scmrun, filter):